Task5/arkpz-pzpi-22-4-ravliuk-kyrylo-task5/Server/CollectionGarbageSystem/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-filter==23.1
matplotlib==3.8.0
numpy==1.26.2
reportlab==4.0.0
psycopg2-binary==2.9.3
psycopg2